Implement installation using rsync (not cp). Should reduce wear of sd card.

+++ pmb/install/format.py
@@ -56,7 +56,8 @@ def format_and_mount_pm_crypt(args):
     mountpoint = "/mnt/install"
     logging.info("(native) format " + device + " (ext4), mount to " +
                  mountpoint)
-    pmb.chroot.root(args, ["mkfs.ext4", "-F", "-q", "-L", "pmOS_root", device])
+    if False:
+        pmb.chroot.root(args, ["mkfs.ext4", "-F", "-q", "-L", "pmOS_root", device])
     pmb.chroot.root(args, ["mkdir", "-p", mountpoint])
     pmb.chroot.root(args, ["mount", device, mountpoint])
